fix: Move external assets copy to postinstall step
